fix: immoDelete is now Removing allImmoImgs from StaticFiles correctly

+++ API/SE_WebAPI/Interfaces/IImgRepository.cs
@@ -6,7 +6,7 @@ namespace SE_WebAPI.Interfaces
     public interface IImgRepository
     {
         List<string> ListarCaminhos(short idImovel);
-        void CadastrarCaminhos(short idImovel,List<string> caminhos);
+        void CadastrarCaminhos(int idImovel,List<string> caminhos);
         void DeletarCaminhos(short idImovel);
     }
 }
